feat: Add PhiView property to Gmodel class
This commit adds a new property called PhiView to the Gmodel class. The PhiView property is used to store a 4x4 matrix representing the view transformation.

The GenerateObject method has been modified to check if the PhiView property is empty before assigning it to the variable Phi. If it is not empty, then the value of Phi is set to the value of the PhiView property.

Additionally, a small bug in the BakeCubemap method has been fixed. The condition `if dY/dX >= 2` has been changed from `I = vertzcat(I,I)` to `I = vertcat(I,I)`. Similarly, `if dX/dY >= 2` has been changed from `I = horzcat(I,I)` to `I = horzcat(I,I)`.